UserPolicy implemented

+++ app/Http/Controllers/UserController.php
@@ -10,13 +10,20 @@ use Illuminate\Http\RedirectResponse;
 
 class UserController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index(): View
     {
+        $this->authorize('viewAny', User::class);
         return view('user.index');
     }
 
     public function create(): View
     {
+        $this->authorize('create', User::class);
         return view('user.create');
     }
 
@@ -27,12 +34,15 @@ class UserController extends Controller
 
     public function show(string $id, UserService $userService): View
     {
+        $user = User::find($id);
+        $this->authorize('view', $user);
         $user = $userService->findById($id);
         return view('user.show', compact('user'));
     }
 
     public function edit(User $user): View
     {
+        $this->authorize('update', $user);
         $id = $user->id;
         return view('user.edit', compact('id'));
     }
@@ -44,6 +54,8 @@ class UserController extends Controller
 
     public function destroy($id, UserService $userService): RedirectResponse
     {
+        $user = User::find($id);
+        $this->authorize('delete', $user);
         $userService->destroyUser($id);
         return redirect()->route('users.index');
           /**
